backend/requirements.txt
fastapi
uvicorn
google-generativeai
google-cloud-firestore
redis
httpx
Pillow
python-dotenv
tenacity
typing_extensions
